refactor: move form and handler tests to dedicated files
